+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - derive `Debug`, `PartialEq`, `Eq` and `Hash` for generated marker types
 - allow passing arguments in `#[marker_type]` invocations that will be applied to the marker type
 - removed `#[variant_amount]` in favor of the new `VariantAmount` derive macro
+- removed `std` feature, the entire crate is `no_std` compatible anyway
 
 ## 0.0.2
 - added `UnwrapVariant` and improved error output